fix: allow for no docroot for Drupal projects
Some projects don't use a docroot/web/public directory so this needs be
able to be set to `null`. As I can't do this with Twig's `is not
defined` test, I'll need to remove the default `web` value and add it to
the build.yaml file for the appropriate projects.

If I can get this to work in the future, I'll re-add `web` as the
default value as that's the default and most common option.