feat(github-actions): run tests first
The tests are the most valuable part of the CI pipeline, so these should
be run first.

If a PHPCS or PHPStan error occurs but the tests pass, we know the
functionality works and the error can be addressed.

If a test fails, the CI pipeline will fail sooner.

In the future, the tasks could be split into separate jobs though it
would need to be done in a way that doesn't add a lot of repetition.

Refs: OD-41